requirements.txt
fastapi
uvicorn[standard]
python-multipart
markdown
jinja2
python-frontmatter
